Splitting behavior in multiple files
