+++ src/core.hpp
@@ -1,0 +1,12 @@
+#ifndef LUASTEAM_CORE
+#define LUASTEAM_CORE
+
+#include "common.hpp"
+
+namespace luasteam {
+
+void add_core(lua_State *L);
+
+} // namespace luasteam
+
+#endif // LUASTEAM_CORE
