+++ src/friends.hpp
@@ -1,0 +1,12 @@
+#ifndef LUASTEAM_FRIENDS
+#define LUASTEAM_FRIENDS
+
+#include "common.hpp"
+
+namespace luasteam {
+
+void add_friends(lua_State *L);
+
+} // namespace luasteam
+
+#endif // LUASTEAM_FRIENDS
