+++ src/user_stats.hpp
@@ -1,0 +1,12 @@
+#ifndef LUASTEAM_USER_STATS
+#define LUASTEAM_USER_STATS
+
+#include "common.hpp"
+
+namespace luasteam {
+
+void add_user_stats(lua_State *L);
+
+} // namespace luasteam
+
+#endif // LUASTEAM_USER_STATS
